+++ CHANGELOG.md
@@ -1,8 +1,25 @@
-# 0.0.1 (May 4, 2021)
+## 0.0.20 (May 6, 2021)
+
+- MeetBox dynamically created inside parent component.
+- Parent component must have some width and height, for example fixed.
+- Connection to PubNub for messaging.
+- Connection to XirSys for relay.
+- Display local stream.
+- Display remote stream.
+- Microphone on/off.
+- Camera on/off.
+- Close connection button.
+- Automatic connection close when other party closed his/her connection.
+- Automatic close when Internet connection fails.
+- Tested to be included in Angular application.
+
+## 0.0.1 (May 4, 2021)
+
 Created initial public repository on GitHub.
 Migrated from private repository of **Engos Software** to public domain.
 
-# 0.0.0 (October 10, 2017)
+## 0.0.0 (October 10, 2017)
+
 Original WebRTC library was developed in Autumn 2017, originally as the project called **kiko.live**.
 It was a proof of concept of using [WebRTC](https://webrtc.org/) technology in day to day communication.
 Because of relatively low demand for such solutions at that time, this project was frozen.